Rename the default providers directory for consistency, and ignore prettier on schemas
