+++ src/default-providers/Anthropic/settings-schema.json
