+++ src/default-providers/ChromeAI/settings-schema.json
