+++ src/default-providers/MistralAI/instructions.ts
