+++ src/default-providers/MistralAI/settings-schema.json
