+++ tsconfig.json
@@ -23,6 +23,6 @@
     "src/*",
     "src/**/*",
     "src/settings/base.json",
-    "src/default_providers/**/*.json"
+    "src/default-providers/**/*.json"
   ]
 }
